refactor: migrate user model imports to schema for improved consistency
- Replaced imports from the deprecated users.model with users.schema across various files, ensuring a unified approach to user data handling.
- Updated related files to reflect the new import paths, enhancing maintainability and clarity in the user module structure.
