+++ db/drop.ts
@@ -1,10 +1,10 @@
 import { exit } from 'node:process';
 import { db } from '@/database.providers';
+import { userFollows, users } from '@/users/users.schema';
 import { articles, favoriteArticles } from '@articles/articles.schema';
 import { comments } from '@comments/comments.schema';
 import dbConfig from '@db/config';
 import { articleTags, tags } from '@tags/tags.schema';
-import { userFollows, users } from '@users/users.model';
 import { getTableName, sql } from 'drizzle-orm';
 
 const tables = [
