+++ src/articles/articles.schema.ts
@@ -1,6 +1,6 @@
+import { users } from '@/users/users.schema';
 import { comments } from '@comments/comments.schema';
 import { articleTags } from '@tags/tags.schema';
-import { users } from '@users/users.model';
 import { relations } from 'drizzle-orm';
 import {
   integer,
