+++ src/comments/comments.schema.ts
@@ -1,5 +1,5 @@
+import { users } from '@/users/users.schema';
 import { articles } from '@articles/articles.schema';
-import { users } from '@users/users.model';
 import { relations } from 'drizzle-orm';
 import { integer, pgTable, serial, text, timestamp } from 'drizzle-orm/pg-core';
