+++ src/database.providers.ts
@@ -1,8 +1,8 @@
+import * as usersSchema from '@/users/users.schema';
 import * as articlesSchema from '@articles/articles.schema';
 import * as commentsSchema from '@comments/comments.schema';
 import { dbCredentialsString } from '@db/config';
 import * as tagsSchema from '@tags/tags.schema';
-import * as usersSchema from '@users/users.model';
 import { drizzle } from 'drizzle-orm/postgres-js';
 import postgres from 'postgres';
